requirements.txt
python-socketio
numpy
websocket-client
